Moved linear slide method into DriveMethods, increased aggresiveness of manual control, changed Teleop down height to 0 clicks, put servoGrabber as universal in Variables, used Alex class to (successfully) demonstrate all auto functions
